Fix ERR (Expanded are shit in Lisaac) and fix tests
I had a gigantic problem: when I tried to clone ERR, I got a compilation
error in C because ERR is expanded (and I think it's perhaps because it
contains Expanded POS).

I tried to get by using a variable that I hoped was auto initialized with
a default value, but then, when I modified this variable, I modified in
fact the prototype ERR.

The solution I found (after trying backticks) is to redefine clone and use
REFERENCE(ERR) to get a clean variable, initialized with a clean expanded
object. And now, error duplication ceases and I could even find a bug when
the correct error appeared but was never linked in the data structure. it
should not have appeared.

I really need to start a new Lysaac compiler to get out of this mess !!!